+++ src/cmd/bucket/show.rs
@@ -97,12 +97,36 @@ fn record_range_cells_compact(oldest: u64, latest: u64, is_empty: bool) -> Vec<S
 }
 
 fn print_bucket(ctx: &CliContext, bucket: FullBucketInfo) -> anyhow::Result<()> {
+    let is_json = ctx.json();
     let info = bucket.info;
     let total_blocks = bucket
         .entries
         .iter()
         .map(|entry| entry.block_count)
         .sum::<u64>();
+
+    let record_data = record_range_cells_compact(
+        info.oldest_record,
+        info.latest_record,
+        info.entry_count == 0,
+    );
+
+    if is_json {
+        let bucket_json = serde_json::json!({
+            "name": info.name,
+            "entries": info.entry_count,
+            "blocks": total_blocks,
+            "size": ByteSize(info.size).display().si().to_string(),
+            "status": format!("{:?}", &info.status),
+            "provisioned": format!("{:?}", info.is_provisioned),
+            "Oldest Record (UTC)": record_data[0].split_once(": ").unwrap().1,
+            "Latest Record (UTC)": record_data[1].split_once(": ").unwrap().1,
+        });
+
+        output!(ctx, "{}", serde_json::to_string(&bucket_json).unwrap());
+        return Ok(());
+    }
+
     let mut info_cells = vec![
         labeled_cell("Name", info.name),
         labeled_cell("Entries", info.entry_count),
@@ -111,11 +135,8 @@ fn print_bucket(ctx: &CliContext, bucket: FullBucketInfo) -> anyhow::Result<()> 
         labeled_cell("Status", print_bucket_status(&info.status)),
         labeled_cell("Provisioned", if info.is_provisioned { "✓" } else { "-" }),
     ];
-    info_cells.extend(record_range_cells_compact(
-        info.oldest_record,
-        info.latest_record,
-        info.entry_count == 0,
-    ));
+
+    info_cells.extend(record_data);
 
     let info_table = build_info_table_with_columns(info_cells, 1);
 
@@ -125,6 +146,14 @@ fn print_bucket(ctx: &CliContext, bucket: FullBucketInfo) -> anyhow::Result<()> 
 }
 
 fn print_full_bucket(ctx: &CliContext, bucket: FullBucketInfo) -> anyhow::Result<()> {
+    let is_json = ctx.json();
+    let full_bucket_info = bucket.clone();
+
+    if is_json {
+        output!(ctx, "{}", serde_json::to_string(&full_bucket_info).unwrap());
+        return Ok(());
+    }
+
     let settings = bucket.settings;
     let info = bucket.info;
     let total_blocks = bucket
@@ -132,6 +161,16 @@ fn print_full_bucket(ctx: &CliContext, bucket: FullBucketInfo) -> anyhow::Result
         .iter()
         .map(|entry| entry.block_count)
         .sum::<u64>();
+
+    let record_data = record_range_cells_compact(
+        info.oldest_record,
+        info.latest_record,
+        info.entry_count == 0,
+    );
+
+    let entries = bucket.entries.into_iter().map(EntryTable::from);
+    let entries: Vec<EntryTable> = entries.collect::<Vec<EntryTable>>();
+
     let mut info_cells = vec![
         labeled_cell("Name", info.name),
         labeled_cell("Quota Type", settings.quota_type.unwrap()),
@@ -158,11 +197,8 @@ fn print_full_bucket(ctx: &CliContext, bucket: FullBucketInfo) -> anyhow::Result
         labeled_cell("Provisioned", if info.is_provisioned { "✓" } else { "-" }),
         String::new(),
     ];
-    for cell in record_range_cells_compact(
-        info.oldest_record,
-        info.latest_record,
-        info.entry_count == 0,
-    ) {
+
+    for cell in record_data {
         info_cells.push(cell);
         info_cells.push(String::new());
     }
@@ -172,7 +208,6 @@ fn print_full_bucket(ctx: &CliContext, bucket: FullBucketInfo) -> anyhow::Result
     output!(ctx, "{}", info_table);
     output!(ctx, "");
 
-    let entries = bucket.entries.into_iter().map(EntryTable::from);
     let table = Table::new(entries).with(Style::markdown()).to_string();
     output!(ctx, "{}", table);
 
@@ -182,7 +217,10 @@ fn print_full_bucket(ctx: &CliContext, bucket: FullBucketInfo) -> anyhow::Result
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::context::tests::{bucket, context};
+    use crate::context::{
+        tests::{bucket, context, MockOutput},
+        ContextBuilder,
+    };
     use reduct_rs::ResourceStatus;
     use rstest::rstest;
 
@@ -316,5 +354,106 @@ mod tests {
         assert_eq!(row.status, "🗑 Deleting");
         assert_eq!(row.oldest_record, "---");
         assert_eq!(row.latest_record, "---");
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_show_bucket_json(context: CliContext, #[future] bucket: String) {
+        let bucket_name = bucket.await;
+        let client = build_client(&context, "local").await.unwrap();
+        client.create_bucket(&bucket_name).send().await.unwrap();
+
+        let args = show_bucket_cmd()
+            .get_matches_from(vec!["show", format!("local/{}", bucket_name).as_str()]);
+
+        let ctx = ContextBuilder::new()
+            .config_path(context.config_path())
+            .json(Some(true))
+            .output(Box::new(MockOutput::new()))
+            .build();
+
+        show_bucket(&ctx, &args).await.unwrap();
+
+        let json: serde_json::Value = serde_json::from_str(&ctx.stdout().history()[0]).unwrap();
+
+        assert_eq!(json["name"], bucket_name);
+        assert_eq!(json["entries"], 0);
+        assert_eq!(json["blocks"], 0);
+        assert_eq!(json["size"], "0 B");
+        assert_eq!(json["status"], "Ready");
+        assert_eq!(json["provisioned"], "false");
+        assert_eq!(json["Oldest Record (UTC)"], "---");
+        assert_eq!(json["Latest Record (UTC)"], "---");
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_show_bucket_full_json(context: CliContext, #[future] bucket: String) {
+        let bucket_name = bucket.await;
+        let client = build_client(&context, "local").await.unwrap();
+        let bucket = client.create_bucket(&bucket_name).send().await.unwrap();
+        bucket
+            .write_record("test")
+            .data("data")
+            .timestamp_us(1)
+            .send()
+            .await
+            .unwrap();
+        bucket
+            .write_record("test")
+            .data("data")
+            .timestamp_us(1000)
+            .send()
+            .await
+            .unwrap();
+
+        let args = show_bucket_cmd().get_matches_from(vec![
+            "show",
+            format!("local/{}", bucket_name).as_str(),
+            "--full",
+        ]);
+
+        let ctx = ContextBuilder::new()
+            .config_path(context.config_path())
+            .json(Some(true))
+            .output(Box::new(MockOutput::new()))
+            .build();
+
+        show_bucket(&ctx, &args).await.unwrap();
+
+        let json: serde_json::Value = serde_json::from_str(&ctx.stdout().history()[0]).unwrap();
+
+        // Verify bucket information
+        let bucket_info = &json["info"];
+        assert_eq!(bucket_info["name"], bucket_name);
+        assert_eq!(bucket_info["entry_count"], serde_json::json!(1));
+        assert_eq!(bucket_info["size"], serde_json::json!(77));
+        assert_eq!(bucket_info["oldest_record"], serde_json::json!(1));
+        assert_eq!(bucket_info["latest_record"], serde_json::json!(1000));
+        assert_eq!(bucket_info["is_provisioned"], serde_json::json!(false));
+        assert_eq!(bucket_info["status"], serde_json::json!("READY"));
+
+        // Verify bucket settings
+        let bucket_settings = &json["settings"];
+        assert_eq!(bucket_settings["quota_type"], serde_json::json!("NONE"));
+        assert_eq!(bucket_settings["quota_size"], serde_json::json!(0));
+        assert_eq!(
+            bucket_settings["max_block_size"],
+            serde_json::json!(64000000)
+        );
+        assert_eq!(
+            bucket_settings["max_block_records"],
+            serde_json::json!(1024)
+        );
+
+        // Verify entries
+        let bucket_entries = &json["entries"];
+        assert_eq!(bucket_entries[0]["name"], serde_json::json!("test"));
+        assert_eq!(bucket_entries[0]["size"], serde_json::json!(77));
+        assert_eq!(bucket_entries[0]["record_count"], serde_json::json!(2));
+        assert_eq!(bucket_entries[0]["block_count"], serde_json::json!(1));
+        assert_eq!(bucket_entries[0]["oldest_record"], serde_json::json!(1));
+        assert_eq!(bucket_entries[0]["latest_record"], serde_json::json!(1000));
+        assert_eq!(bucket_entries[0]["status"], serde_json::json!("READY"));
     }
 }